Cache the farthest back we have retrieved posts
This should reduce danbooru pulls when there are no new posts from
n (how many pages we need to go back to find a new post,) to 2 if there
are still good posts on the page, or 3 if we need a new page.

This is a good compromise until I decide how I want to cache the
danbooru search results entirely.